Update version to 1.3.1 and adjust artifact name
Adjusted #versionReleaseNotes styling for improved usability by setting its height to 90vh with auto overflow and a customized scrollbar for a cleaner look. Set the parent container height to 100% for better layout consistency. Also bumped the application version from 1.3.0 to 1.3.1 in package.json and package-lock.json, and updated the artifactName in electron-builder.yml to consistently use the PlayClan-Launcher prefix.